+++ web/apps/photos/src/utils/file/index.ts
@@ -1,6 +1,4 @@
 import { FILE_TYPE } from "@/media/file-type";
-import { isNonWebImageFileExtension } from "@/media/formats";
-import { heicToJPEG } from "@/media/heic-convert";
 import { decodeLivePhoto } from "@/media/live-photo";
 import {
     EncryptedEnteFile,
@@ -12,20 +10,18 @@ import {
     FileWithUpdatedMagicMetadata,
 } from "@/new/photos/types/file";
 import { VISIBILITY_STATE } from "@/new/photos/types/magicMetadata";
+import { detectFileTypeInfo } from "@/new/photos/utils/detect-type";
 import { mergeMetadata } from "@/new/photos/utils/file";
 import { lowercaseExtension } from "@/next/file";
 import log from "@/next/log";
-import { CustomErrorMessage, type Electron } from "@/next/types/ipc";
-import { workerBridge } from "@/next/worker/worker-bridge";
+import { type Electron } from "@/next/types/ipc";
 import { withTimeout } from "@/utils/promise";
 import ComlinkCryptoWorker from "@ente/shared/crypto";
 import { LS_KEYS, getData } from "@ente/shared/storage/localStorage";
 import type { User } from "@ente/shared/user/types";
 import { downloadUsingAnchor } from "@ente/shared/utils";
 import { t } from "i18next";
-import isElectron from "is-electron";
 import { moveToHiddenCollection } from "services/collectionService";
-import { detectFileTypeInfo } from "services/detect-type";
 import DownloadManager from "services/download";
 import { updateFileCreationDateInEXIF } from "services/exif";
 import {
@@ -43,19 +39,6 @@ import { isArchivedFile, updateMagicMetadata } from "utils/magicMetadata";
 import { safeFileName } from "utils/native-fs";
 import { writeStream } from "utils/native-stream";
 
-const SUPPORTED_RAW_FORMATS = [
-    "heic",
-    "rw2",
-    "tiff",
-    "arw",
-    "cr3",
-    "cr2",
-    "nef",
-    "psd",
-    "dng",
-    "tif",
-];
-
 export enum FILE_OPS_TYPE {
     DOWNLOAD,
     FIX_TIME,
@@ -65,22 +48,6 @@ export enum FILE_OPS_TYPE {
     TRASH,
     DELETE_PERMANENTLY,
 }
-
-class ModuleState {
-    /**
-     * This will be set to true if we get an error from the Node.js side of our
-     * desktop app telling us that native JPEG conversion is not available for
-     * the current OS/arch combination.
-     *
-     * That way, we can stop pestering it again and again (saving an IPC
-     * round-trip).
-     *
-     * Note the double negative when it is used.
-     */
-    isNativeJPEGConversionNotAvailable = false;
-}
-
-const moduleState = new ModuleState();
 
 /**
  * @returns a string to use as an identifier when logging information about the
@@ -255,80 +222,6 @@ export async function decryptFile(
         log.error("file decryption failed", e);
         throw e;
     }
-}
-
-/**
- * The returned blob.type is filled in, whenever possible, with the MIME type of
- * the data that we're dealing with.
- */
-export const getRenderableImage = async (fileName: string, imageBlob: Blob) => {
-    try {
-        const tempFile = new File([imageBlob], fileName);
-        const fileTypeInfo = await detectFileTypeInfo(tempFile);
-        log.debug(
-            () =>
-                `Need renderable image for ${JSON.stringify({ fileName, ...fileTypeInfo })}`,
-        );
-        const { extension } = fileTypeInfo;
-
-        if (!isNonWebImageFileExtension(extension)) {
-            // Either it is something that the browser already knows how to
-            // render, or something we don't even about yet.
-            const mimeType = fileTypeInfo.mimeType;
-            if (!mimeType) {
-                log.info(
-                    "Trying to render a file without a MIME type",
-                    fileName,
-                );
-                return imageBlob;
-            } else {
-                return new Blob([imageBlob], { type: mimeType });
-            }
-        }
-
-        const available = !moduleState.isNativeJPEGConversionNotAvailable;
-        if (isElectron() && available && isSupportedRawFormat(extension)) {
-            // If we're running in our desktop app, see if our Node.js layer can
-            // convert this into a JPEG using native tools for us.
-            try {
-                return await nativeConvertToJPEG(imageBlob);
-            } catch (e) {
-                if (e.message.endsWith(CustomErrorMessage.NotAvailable)) {
-                    moduleState.isNativeJPEGConversionNotAvailable = true;
-                } else {
-                    log.error("Native conversion to JPEG failed", e);
-                }
-            }
-        }
-
-        if (extension == "heic" || extension == "heif") {
-            // For HEIC/HEIF files we can use our web HEIC converter.
-            return await heicToJPEG(imageBlob);
-        }
-
-        return undefined;
-    } catch (e) {
-        log.error(`Failed to get renderable image for ${fileName}`, e);
-        return undefined;
-    }
-};
-
-const nativeConvertToJPEG = async (imageBlob: Blob) => {
-    const startTime = Date.now();
-    const imageData = new Uint8Array(await imageBlob.arrayBuffer());
-    const electron = globalThis.electron;
-    // If we're running in a worker, we need to reroute the request back to
-    // the main thread since workers don't have access to the `window` (and
-    // thus, to the `window.electron`) object.
-    const jpegData = electron
-        ? await electron.convertToJPEG(imageData)
-        : await workerBridge.convertToJPEG(imageData);
-    log.debug(() => `Native JPEG conversion took ${Date.now() - startTime} ms`);
-    return new Blob([jpegData], { type: "image/jpeg" });
-};
-
-export function isSupportedRawFormat(exactType: string) {
-    return SUPPORTED_RAW_FORMATS.includes(exactType.toLowerCase());
 }
 
 export async function changeFilesVisibility(
