+++ web/packages/new/photos/utils/file.ts
@@ -1,4 +1,40 @@
+import { isNonWebImageFileExtension } from "@/media/formats";
+import { heicToJPEG } from "@/media/heic-convert";
+import log from "@/next/log";
+import { CustomErrorMessage } from "@/next/types/ipc";
+import { workerBridge } from "@/next/worker/worker-bridge";
+import isElectron from "is-electron";
 import type { EnteFile } from "../types/file";
+import { detectFileTypeInfo } from "./detect-type";
+
+const SUPPORTED_RAW_FORMATS = [
+    "heic",
+    "rw2",
+    "tiff",
+    "arw",
+    "cr3",
+    "cr2",
+    "nef",
+    "psd",
+    "dng",
+    "tif",
+];
+
+class ModuleState {
+    /**
+     * This will be set to true if we get an error from the Node.js side of our
+     * desktop app telling us that native JPEG conversion is not available for
+     * the current OS/arch combination.
+     *
+     * That way, we can stop pestering it again and again (saving an IPC
+     * round-trip).
+     *
+     * Note the double negative when it is used.
+     */
+    isNativeJPEGConversionNotAvailable = false;
+}
+
+const moduleState = new ModuleState();
 
 /**
  * [Note: File name for local EnteFile objects]
@@ -27,4 +63,81 @@ export function mergeMetadata(files: EnteFile[]): EnteFile[] {
 
         return file;
     });
+}
+
+/**
+ * The returned blob.type is filled in, whenever possible, with the MIME type of
+ * the data that we're dealing with.
+ */
+export const getRenderableImage = async (fileName: string, imageBlob: Blob) => {
+    try {
+        const tempFile = new File([imageBlob], fileName);
+        const fileTypeInfo = await detectFileTypeInfo(tempFile);
+        log.debug(
+            () =>
+                `Need renderable image for ${JSON.stringify({ fileName, ...fileTypeInfo })}`,
+        );
+        const { extension } = fileTypeInfo;
+
+        if (!isNonWebImageFileExtension(extension)) {
+            // Either it is something that the browser already knows how to
+            // render, or something we don't even about yet.
+            const mimeType = fileTypeInfo.mimeType;
+            if (!mimeType) {
+                log.info(
+                    "Trying to render a file without a MIME type",
+                    fileName,
+                );
+                return imageBlob;
+            } else {
+                return new Blob([imageBlob], { type: mimeType });
+            }
+        }
+
+        const available = !moduleState.isNativeJPEGConversionNotAvailable;
+        if (isElectron() && available && isSupportedRawFormat(extension)) {
+            // If we're running in our desktop app, see if our Node.js layer can
+            // convert this into a JPEG using native tools for us.
+            try {
+                return await nativeConvertToJPEG(imageBlob);
+            } catch (e) {
+                if (
+                    e instanceof Error &&
+                    e.message.endsWith(CustomErrorMessage.NotAvailable)
+                ) {
+                    moduleState.isNativeJPEGConversionNotAvailable = true;
+                } else {
+                    log.error("Native conversion to JPEG failed", e);
+                }
+            }
+        }
+
+        if (extension == "heic" || extension == "heif") {
+            // For HEIC/HEIF files we can use our web HEIC converter.
+            return await heicToJPEG(imageBlob);
+        }
+
+        return undefined;
+    } catch (e) {
+        log.error(`Failed to get renderable image for ${fileName}`, e);
+        return undefined;
+    }
+};
+
+const nativeConvertToJPEG = async (imageBlob: Blob) => {
+    const startTime = Date.now();
+    const imageData = new Uint8Array(await imageBlob.arrayBuffer());
+    const electron = globalThis.electron;
+    // If we're running in a worker, we need to reroute the request back to
+    // the main thread since workers don't have access to the `window` (and
+    // thus, to the `window.electron`) object.
+    const jpegData = electron
+        ? await electron.convertToJPEG(imageData)
+        : await workerBridge.convertToJPEG(imageData);
+    log.debug(() => `Native JPEG conversion took ${Date.now() - startTime} ms`);
+    return new Blob([jpegData], { type: "image/jpeg" });
+};
+
+export function isSupportedRawFormat(exactType: string) {
+    return SUPPORTED_RAW_FORMATS.includes(exactType.toLowerCase());
 }
